requirements.txt
PyExecJS
requests
loguru
# pip install -r requirements.txt -i https://pypi.mirrors.ustc.edu.cn/simple/